+++ laktory/models/dataquality/__init__.py
@@ -1,0 +1,2 @@
+from laktory.models.dataquality.expectation import DataQualityExpectation
+from laktory.models.dataquality.check import DataQualityCheck